requirements.txt
pandas
dotenv
google-generativeai
